requirements.txt
transformers~=4.21
bitsandbytes~=0.35
omegaconf~=2.2
pytorch-lightning~=1.8
numpy~=1.21
Pillow~=9.3
tqdm~=4.64
click~=8.1
ftfy~=6.1
accelerate~=0.14
psutil~=5.9
requests~=2.28
diffusers~=0.11
torch-ema==0.3
safetensors~=0.2